+++ valhalla/thor/costmatrix.h
@@ -139,8 +139,8 @@ protected:
   std::array<uint32_t, 2> locs_count_;
   std::array<uint32_t, 2> locs_remaining_;
 
-  // The cost threshold being used for the currently executing query
-  float current_cost_threshold_;
+  // The ptth distance threshold being used for the currently executing query
+  float current_pathdist_threshold_;
 
   // Status
   std::array<std::vector<LocationStatus>, 2> locs_status_;
@@ -160,11 +160,11 @@ protected:
   baldr::DateTime::tz_sys_info_cache_t tz_cache_;
 
   /**
-   * Get the cost threshold based on the current mode and the max arc-length distance
-   * for that mode.
+   * Get the cost range for the doublebucket queue based on the current mode and the max arc-length
+   * distance for that mode.
    * @param  max_matrix_distance   Maximum arc-length distance for current mode.
    */
-  float GetCostThreshold(const float max_matrix_distance);
+  float GetBucketRange(const float max_matrix_distance);
 
   /**
    * Form the initial time distance matrix given the sources